backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "IlyaVafin",email = "rthrthrer5gh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)"
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
